requirements.txt
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
asttokens==3.0.0
attrs==24.2.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
black==24.10.0
blinker==1.9.0
blis==1.0.1
bm25s==0.2.4
build==1.2.2.post1
cachetools==5.5.0
catalogue==2.0.10
cbor==1.0.0
certifi==2024.8.30
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.5.18
click==8.1.7
cloudpathlib==0.20.0
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.5
cymem==2.0.8
debugpy==1.8.9
decorator==5.1.1
Deprecated==1.2.15
durationpy==0.9
es_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-3.8.0/es_core_news_sm-3.8.0-py3-none-any.whl
exceptiongroup==1.2.2
executing==2.1.0
faiss-cpu==1.9.0
fastapi==0.115.5
filelock==3.16.1
flatbuffers==24.3.25
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.36.0
googleapis-common-protos==1.66.0
grpcio==1.68.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.10
ijson==3.3.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
inscriptis==2.5.0
ipykernel==6.29.5
ipython==8.30.0
ir_datasets==0.5.9
jax==0.4.35
jaxlib==0.4.35
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kubernetes==31.0.0
langcodes==3.5.0
language_data==1.3.0
llvmlite==0.43.0
lxml==5.3.0
lz4==4.3.3
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.0
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
murmurhash==1.0.10
mypy-extensions==1.0.0
narwhals==1.13.3
nest-asyncio==1.6.0
nltk==3.9.1
numba==0.60.0
numpy==2.0.2
oauthlib==3.2.2
onnxruntime==1.20.0
opentelemetry-api==1.28.1
opentelemetry-exporter-otlp-proto-common==1.28.1
opentelemetry-exporter-otlp-proto-grpc==1.28.1
opentelemetry-instrumentation==0.49b1
opentelemetry-instrumentation-asgi==0.49b1
opentelemetry-instrumentation-fastapi==0.49b1
opentelemetry-proto==1.28.1
opentelemetry-sdk==1.28.1
opentelemetry-semantic-conventions==0.49b1
opentelemetry-util-http==0.49b1
opt_einsum==3.4.0
orjson==3.10.11
overrides==7.7.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
posthog==3.7.0
preshed==3.0.9
prompt_toolkit==3.0.48
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
PyPika==0.48.9
pyproject_hooks==1.2.0
PyStemmer==2.2.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytrec-eval==0.5
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.21.0
rsa==4.9
scipy==1.14.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
starlette==0.41.2
streamlit==1.40.0
streamlit-code-editor==0.1.21
sympy==1.13.3
tenacity==9.0.0
thinc==8.3.2
tokenizers==0.20.3
toml==0.10.2
tomli==2.1.0
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
trec-car-tools==2.6
typer==0.13.0
typing_extensions==4.12.2
tzdata==2024.2
unlzw3==0.2.2
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
wasabi==1.1.3
watchdog==5.0.3
watchfiles==0.24.0
wcwidth==0.2.13
weasel==0.4.1
websocket-client==1.8.0
websockets==14.1
wrapt==1.16.0
zipp==3.21.0
zlib-state==0.1.9
